Fixing 42 pc warnings

+++ ex00/ft_utils.cpp
@@ -1,4 +1,14 @@
-// CABECERA
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_utils.cpp                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: pabad-ap <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/01/28 17:38:06 by pabad-ap          #+#    #+#             */
+/*   Updated: 2025/01/28 17:40:13 by pabad-ap         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "ft_utils.hpp"
 
@@ -14,7 +24,7 @@ std::string		ft_trim( std::string str )
 // Remove leading isspace characters in a string
 static	std::string	ft_ltrim( std::string str )
 {
-	int	i = 0;
+	size_t	i = 0;
 	
 	for ( std::string::iterator it = str.begin(); it != str.end(); it++ )
 	{
@@ -29,7 +39,7 @@ static	std::string	ft_ltrim( std::string str )
 // Remove trailing isspace characters in a string
 static	std::string	ft_rtrim( std::string str )
 {
-	int	i = 0;
+	size_t	i = 0;
 	
 	for ( std::string::reverse_iterator it = str.rbegin(); it != str.rend(); ++it )
 	{
